Changes to the Folder Structure (on server side), modularized with routers for AWS and User Info
