+++ server/controllers/aws/Credentials/libs/stsClient.js
